Move Todo and configuration model in to Relevant folders
